moved to subcomponent, cleaned up the view
