+++ src/components.d.ts
@@ -12,8 +12,8 @@ declare module 'vue' {
     AddCompanyDialog: typeof import('./components/dialog/AddCompanyDialog.vue')['default']
     AddEmployedDialog: typeof import('./components/dialog/AddEmployedDialog.vue')['default']
     AddSuggestionDialog: typeof import('./components/dialog/AddSuggestionDialog.vue')['default']
+    ArticleCard: typeof import('./components/cards/ArticleCard.vue')['default']
     EntityPicker: typeof import('./components/forms/EntityPicker.vue')['default']
-    HelpView: typeof import('./components/HelpView.vue')['default']
     Home: typeof import('./components/Home.vue')['default']
     HomeItem: typeof import('./components/HomeItem.vue')['default']
     ListView: typeof import('./components/ListView.vue')['default']
